keymaps: unsane the next/previous bindings
this is something in lazyvim that someone thought would make things less
confusing, but I depend on the default behaviour as it makes a lot more
sense to me.

this means that if you search for a word under the cursor using `#`,
then `n` will move _up_ instead of the "normal" direction down.